requirements.txt
anthropic>=0.34.0
openai>=1.40.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.1.0
moviepy>=1.0.3
pydub>=0.25.1
requests>=2.31.0
python-dotenv>=1.0.0
mutagen>=1.47.0
Pillow>=10.0.0
